+++ backend/src/baserow/contrib/database/api/views/grid/views.py
@@ -1,3 +1,4 @@
+from django.db import transaction
 from drf_spectacular.openapi import OpenApiParameter, OpenApiTypes
 from drf_spectacular.utils import extend_schema
 from rest_framework.pagination import LimitOffsetPagination
@@ -18,10 +19,14 @@ from baserow.contrib.database.api.rows.serializers import (
     get_row_serializer_class,
     RowSerializer,
 )
+from baserow.contrib.database.api.views.errors import ERROR_VIEW_DOES_NOT_EXIST
 from baserow.contrib.database.api.views.grid.serializers import (
     GridViewFieldOptionsSerializer,
+    PublicGridViewInfoSerializer,
 )
-from baserow.contrib.database.api.views.serializers import FieldOptionsField
+from baserow.contrib.database.api.views.serializers import (
+    FieldOptionsField,
+)
 from baserow.contrib.database.rows.registries import row_metadata_registry
 from baserow.contrib.database.views.exceptions import ViewDoesNotExist
 from baserow.contrib.database.views.handler import ViewHandler
@@ -259,3 +264,215 @@ class GridViewView(APIView):
         )
         serializer = serializer_class(results, many=True)
         return Response(serializer.data)
+
+
+class PublicGridViewRowsView(APIView):
+    permission_classes = (AllowAny,)
+
+    @extend_schema(
+        parameters=[
+            OpenApiParameter(
+                name="slug",
+                location=OpenApiParameter.PATH,
+                type=OpenApiTypes.STR,
+                description="Returns only rows that belong to the related view's "
+                "table.",
+            ),
+            OpenApiParameter(
+                name="count",
+                location=OpenApiParameter.QUERY,
+                type=OpenApiTypes.NONE,
+                description="If provided only the count will be returned.",
+            ),
+            OpenApiParameter(
+                name="include",
+                location=OpenApiParameter.QUERY,
+                type=OpenApiTypes.STR,
+                description=(
+                    "A comma separated list allowing the values of "
+                    "`field_options` which will add the object/objects with the "
+                    "same "
+                    "name to the response if included. The `field_options` object "
+                    "contains user defined view settings for each field. For "
+                    "example the field's width is included in here."
+                ),
+            ),
+            OpenApiParameter(
+                name="limit",
+                location=OpenApiParameter.QUERY,
+                type=OpenApiTypes.INT,
+                description="Defines how many rows should be returned.",
+            ),
+            OpenApiParameter(
+                name="offset",
+                location=OpenApiParameter.QUERY,
+                type=OpenApiTypes.INT,
+                description="Can only be used in combination with the `limit` "
+                "parameter and defines from which offset the rows should "
+                "be returned.",
+            ),
+            OpenApiParameter(
+                name="page",
+                location=OpenApiParameter.QUERY,
+                type=OpenApiTypes.INT,
+                description="Defines which page of rows should be returned. Either "
+                "the `page` or `limit` can be provided, not both.",
+            ),
+            OpenApiParameter(
+                name="size",
+                location=OpenApiParameter.QUERY,
+                type=OpenApiTypes.INT,
+                description="Can only be used in combination with the `page` parameter "
+                "and defines how many rows should be returned.",
+            ),
+            OpenApiParameter(
+                name="search",
+                location=OpenApiParameter.QUERY,
+                type=OpenApiTypes.STR,
+                description="If provided only rows with data that matches the search "
+                "query are going to be returned.",
+            ),
+        ],
+        tags=["Database table grid view"],
+        operation_id="public_list_database_table_grid_view_rows",
+        description=(
+            "Lists the requested rows of the view's table related to the provided "
+            "`slug` if the grid view is public."
+            "The response is paginated either by a limit/offset or page/size style. "
+            "The style depends on the provided GET parameters. The properties of the "
+            "returned rows depends on which fields the table has. For a complete "
+            "overview of fields use the **list_database_table_fields** endpoint to "
+            "list them all. In the example all field types are listed, but normally "
+            "the number in field_{id} key is going to be the id of the field. "
+            "The value is what the user has provided and the format of it depends on "
+            "the fields type.\n"
+            "\n"
+        ),
+        responses={
+            200: get_example_pagination_serializer_class(
+                get_example_row_serializer_class(add_id=True, user_field_names=False),
+                additional_fields={
+                    "field_options": FieldOptionsField(
+                        serializer_class=GridViewFieldOptionsSerializer, required=False
+                    ),
+                },
+                serializer_name="PublicPaginationSerializerWithGridViewFieldOptions",
+            ),
+            400: get_error_schema(["ERROR_USER_NOT_IN_GROUP"]),
+            404: get_error_schema(["ERROR_GRID_DOES_NOT_EXIST"]),
+        },
+    )
+    @map_exceptions(
+        {
+            UserNotInGroup: ERROR_USER_NOT_IN_GROUP,
+            ViewDoesNotExist: ERROR_GRID_DOES_NOT_EXIST,
+        }
+    )
+    @allowed_includes("field_options")
+    def get(self, request, slug, field_options):
+        """
+        Lists all the rows of a grid view, paginated either by a page or offset/limit.
+        If the limit get parameter is provided the limit/offset pagination will be used
+        else the page number pagination.
+
+        Optionally the field options can also be included in the response if the the
+        `field_options` are provided in the include GET parameter.
+        """
+
+        search = request.GET.get("search")
+
+        view_handler = ViewHandler()
+        view = view_handler.get_public_view_by_slug(request.user, slug, GridView)
+        view_type = view_type_registry.get_by_model(view)
+
+        publicly_visible_field_options = view_type.get_visible_field_options_in_order(
+            view
+        )
+        publicly_visible_field_ids = {
+            o.field_id for o in publicly_visible_field_options
+        }
+
+        # We have to still make a model with all fields as the public rows should still
+        # be filtered by hidden fields.
+        model = view.table.get_model()
+        queryset = view_handler.get_queryset(
+            view,
+            search,
+            model,
+            only_sort_by_field_ids=publicly_visible_field_ids,
+            only_search_by_field_ids=publicly_visible_field_ids,
+        )
+
+        if "count" in request.GET:
+            return Response({"count": queryset.count()})
+
+        if LimitOffsetPagination.limit_query_param in request.GET:
+            paginator = LimitOffsetPagination()
+        else:
+            paginator = PageNumberPagination()
+
+        page = paginator.paginate_queryset(queryset, request, self)
+        serializer_class = get_row_serializer_class(
+            model, RowSerializer, is_response=True, field_ids=publicly_visible_field_ids
+        )
+        serializer = serializer_class(page, many=True)
+
+        response = paginator.get_paginated_response(serializer.data)
+
+        if field_options:
+            context = {"field_options": publicly_visible_field_options}
+            serializer_class = view_type.get_field_options_serializer_class(
+                create_if_missing=False
+            )
+            response.data.update(**serializer_class(view, context=context).data)
+
+        return response
+
+
+class PublicGridViewInfoView(APIView):
+    permission_classes = (AllowAny,)
+
+    @extend_schema(
+        parameters=[
+            OpenApiParameter(
+                name="slug",
+                location=OpenApiParameter.PATH,
+                type=OpenApiTypes.STR,
+                required=True,
+                description="The slug of the grid view to get public information "
+                "about.",
+            )
+        ],
+        tags=["Database table grid view"],
+        operation_id="get_public_grid_view_info",
+        description=(
+            "Returns the required public information to display a single "
+            "shared grid view."
+        ),
+        request=None,
+        responses={
+            200: PublicGridViewInfoSerializer,
+            400: get_error_schema(["ERROR_USER_NOT_IN_GROUP"]),
+            404: get_error_schema(["ERROR_VIEW_DOES_NOT_EXIST"]),
+        },
+    )
+    @map_exceptions(
+        {
+            UserNotInGroup: ERROR_USER_NOT_IN_GROUP,
+            ViewDoesNotExist: ERROR_VIEW_DOES_NOT_EXIST,
+        }
+    )
+    @transaction.atomic
+    def get(self, request, slug):
+
+        handler = ViewHandler()
+        view = handler.get_public_view_by_slug(request.user, slug, view_model=GridView)
+        grid_view_type = view_type_registry.get_by_model(view)
+        field_options = grid_view_type.get_visible_field_options_in_order(view)
+
+        return Response(
+            PublicGridViewInfoSerializer(
+                view=view,
+                fields=[o.field for o in field_options.select_related("field")],
+            ).data
+        )