+++ backend/src/baserow/contrib/database/api/views/serializers.py
@@ -78,15 +78,17 @@ class FieldOptionsField(serializers.Field):
         """
 
         if isinstance(value, View):
-            # If the fields are in the context we can pass them into the
-            # `get_field_options` call so that they don't have to be fetched from the
-            # database again.
-            fields = self.context.get("fields")
+            field_options = self.context.get("field_options", None)
+            if field_options is None:
+                # If the fields are in the context we can pass them into the
+                # `get_field_options` call so that they don't have to be fetched from
+                # the database again.
+                field_options = value.get_field_options(
+                    self.create_if_missing, self.context.get("fields")
+                )
             return {
                 field_options.field_id: self.serializer_class(field_options).data
-                for field_options in value.get_field_options(
-                    self.create_if_missing, fields
-                )
+                for field_options in field_options
             }
         else:
             return value