+++ web-frontend/test/unit/database/table.spec.js
@@ -123,7 +123,7 @@ describe('Table Component Tests', () => {
   async function givenASingleSimpleTableInTheServer() {
     const table = mockServer.createTable()
     const { application } = await mockServer.createAppAndGroup(table)
-    const gridView = mockServer.createGridView(application, table)
+    const gridView = mockServer.createGridView(application, table, {})
     const fields = mockServer.createFields(application, table, [
       {
         name: 'Name',