+++ changelog.md
@@ -15,6 +15,7 @@
   `/database/views/${viewId}/rotate-slug/`.
 * Increased maximum length of application name to 160 characters.
 * Fixed copying/pasting for date field.
+* Added ability to share grid views publicly.
 * Allow changing the text of the submit button in the form view.
 * Fixed reordering of single select options when initially creating the field.
 * Improved performance by not rendering cells that are out of the view port.